Organize commands to subpackages based on their type.
