+++ src/main/java/seedu/address/logic/commands/tagcommands/ListTagsCommand.java
@@ -1,0 +1,5 @@
+package seedu.address.logic.commands.tagcommands;
+
+public class ListTagsCommand {
+    //To be implemented
+}
